app/requirements.txt
BeautifulSoup4
newspaper3k